basic frontend ui layout
